backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
openai==1.3.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1 